Add store/v2 (#3832)
* Use core/v2 and core/v3 modules
* Update TypeMeta to include encoding and compression fields
* Remove ObjectMeta from store.Wrapper

Also remove the automatically generated tests for metadata. These
tests don't seem to get generated correctly, and fail sporadically
when enums are used.

* Move things around
* Move encoding and compression to wrapper
* Move wrapper into the wrap package
* store.WrapResource -> wrap.Resource

* Implement store/v2

This commit implements store/v2, and an etcdstore within it. It
builds upon the wrapper concept introduced to the store in earlier
commits.

Some exploratory work and benchmarks have been written to support
this work. They show that the new store is more or less in line
with the performance of the old store, even though we are now doing
double-encoding and double-decoding of protobufs.

Some changes to the type system have been implemented. Particularly,
the types package has gained the ResolveRaw function, which can
resolve either core/v2 or core/v3 resources. There is also a
RegisterResolver function that can be used with any data type,
instead of just with core/v2.Resource.

Signed-off-by: Eric Chlebek <eric@sensu.io>