Add Scheduler field to the core/v2 Check and CheckConfig APIs (#4089)
* Add the 'Scheduler' field to CheckConfig and Check.

The 'Scheduler' field is a property set by Sensu to indicate the kind of
scheduling the check is expected to undergo. It does not control the
scheduler and is only an informational property. Users should not set
Scheduler.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Integrate check.Scheduler field

* Add to Keepalived
* Add to the check store
* Add to check watchers

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Update CHANGELOG.md

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Use the check Scheduler field as a hint for scheduler restart

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Log ring entity timeout

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Unconditionally cancel ring watchers

This commit causes ring watchers to cancel unconditionally when rings
are updated.

Previously, the scheduler tried to be optimal about this, avoiding
cancelling any ring watchers if the perceived schedule had not changed.
However, now that we have pluggable schedulers, this logic is faulty and
must be removed.

This should not change the behaviour of round-robin checks for OSS and
may even improve the situation for backends that are experiencing issues
with round-robin scheduling.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Fix race condition between schedulerd and ring

Signed-off-by: Eric Chlebek <eric@sensu.io>