Store output in check history
Since only a few history entries are kept, the history size doesn't grow much.
On the positive side, this can be used for the checks to give more information downstream, about their work.

Alternatively we could consider a more complex change adding a boolean option to CheckConfig/Check, and storing the output only if that is set.

Protocol buffer compilation result is added as a followup commit.

Signed-off-by: Guido Trotter <guido@hudson-trading.com>